collector_service/requirements.txt
flask
requests
exchange-calendars
google-cloud-storage

